--- src/operators/mod.rs
@@ -1,7 +1,0 @@
-pub use composite_operators::*;
-pub use math_operators::*;
-pub use stack_operators::*;
-
-mod composite_operators;
-mod math_operators;
-mod stack_operators;
